143: GlyphPointF

+++ FontRasterizerSample.WinForms/GlyphPathBuilderVxs.cs
@@ -29,19 +29,19 @@ namespace PixelFarm.Agg
         {
             ps.CloseFigure();
         }
-        protected override void OnCurve3(short p2x, short p2y, short x, short y)
+        protected override void OnCurve3(float p2x, float p2y, float x, float y)
         {
             ps.Curve3(p2x, p2y, x, y);
         }
-        protected override void OnCurve4(short p2x, short p2y, short p3x, short p3y, short x, short y)
+        protected override void OnCurve4(float p2x, float p2y, float p3x, float p3y, float x, float y)
         {
             ps.Curve4(p2x, p2y, p3x, p3y, x, y);
         }
-        protected override void OnLineTo(short x, short y)
+        protected override void OnLineTo(float x, float y)
         {
             ps.LineTo(x, y);
         }
-        protected override void OnMoveTo(short x, short y)
+        protected override void OnMoveTo(float x, float y)
         {
             ps.MoveTo(x, y);
         }
